pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mira"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "mira-sdk>=0.1.14",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
]